fix: populate  process.env at module load time
BREAKING CHANGE: importing @carforyou/configuration loads the configuration as a side effect, making it accessible immediately. please remove any calls to `loadConfiguration()` and consult the updated README on the new usage
