+++ package.json
@@ -18,6 +18,7 @@
   "license": "MIT",
   "homepage": "https://github.com/carforyou/configuration-pkg#readme",
   "devDependencies": {
+    "@babel/plugin-transform-modules-commonjs": "^7.10.4",
     "@carforyou/eslint-config": "1.2.0",
     "@pika/pack": "^0.5.0",
     "@pika/plugin-build-node": "^0.9.0",
